+++ R/mice.impute.2l.bin.R
@@ -24,8 +24,8 @@
 #'@examples
 #'library(tidyr)
 #'library(dplyr)
-#'data("toenail", package = "HSAUR3")
-#'data <- tidyr::complete(toenail, patientID, visit) %>% 
+#'data("toenail2")
+#'data <- tidyr::complete(toenail2, patientID, visit) %>% 
 #'  tidyr::fill(treatment) %>% 
 #'  dplyr::select(-time) %>%
 #'  dplyr::mutate(patientID = as.integer(patientID))