+++ data-raw/R/toenail2.R
@@ -1,0 +1,3 @@
+data("toenail", package = "HSAUR3")
+toenail2 <- toenail
+usethis::use_data(toenail2, overwrite = TRUE)